Use `short' values for init_pair()ing.
git-svn-id: file:///Users/kling/src/protoman/svn/trunk/ytalk@537 8120b729-f3fe-0310-9465-b9035d29392c

+++ src/cwin.c
@@ -236,7 +236,7 @@ static void
 curses_start()
 {
 #ifdef YTALK_COLOR
-	int i, fg, bg;
+	short i, fg, bg;
 #endif
 	/*
 	 * This little snippet is currently on probation. Although the
